Fix trace audit structure
